+++ core-ng/src/main/java/core/framework/internal/web/route/URLHandler.java
@@ -1,9 +1,9 @@
 package core.framework.internal.web.route;
 
-import core.framework.api.http.HTTPStatus;
 import core.framework.http.HTTPMethod;
 import core.framework.internal.web.controller.ControllerHolder;
 import core.framework.util.Maps;
+import core.framework.web.exception.MethodNotAllowedException;
 
 import java.util.Map;
 
@@ -12,7 +12,7 @@ import static core.framework.util.Strings.format;
 /**
  * @author neo
  */
-public class URLHandler {
+class URLHandler {
     final String pathPattern;
     private final Map<HTTPMethod, ControllerHolder> controllers = Maps.newEnumMap(HTTPMethod.class);
 
@@ -27,11 +27,11 @@ public class URLHandler {
         }
     }
 
-    Route.RouteResult get(HTTPMethod method) {
+    ControllerHolder get(HTTPMethod method) {
         ControllerHolder controller = controllers.get(method);
         if (controller == null) {
-            return new Route.RouteResult(null, HTTPStatus.METHOD_NOT_ALLOWED);
+            throw new MethodNotAllowedException("method not allowed, method=" + method);
         }
-        return new Route.RouteResult(controller, null);
+        return controller;
     }
 }