+++ core-ng/src/main/java/core/framework/internal/web/session/SessionStore.java
@@ -1,5 +1,7 @@
 package core.framework.internal.web.session;
 
+import org.jspecify.annotations.Nullable;
+
 import java.time.Duration;
 import java.util.Map;
 import java.util.Set;
@@ -8,9 +10,10 @@ import java.util.Set;
  * @author neo
  */
 public interface SessionStore {
+    @Nullable
     Map<String, String> getAndRefresh(String sessionId, String domain, Duration timeout);
 
-    void save(String sessionId, String domain, Map<String, String> values, Set<String> changedFields, Duration timeout);
+    void save(String sessionId, String domain, Map<String, @Nullable String> values, Set<String> changedFields, Duration timeout);
 
     void invalidate(String sessionId, String domain);