+++ core-ng/src/main/java/core/framework/internal/web/sse/ServerSentEventHandler.java
@@ -1,6 +1,5 @@
 package core.framework.internal.web.sse;
 
-import core.framework.api.http.HTTPStatus;
 import core.framework.http.HTTPMethod;
 import core.framework.internal.async.VirtualThread;
 import core.framework.internal.log.ActionLog;
@@ -12,12 +11,17 @@ import core.framework.internal.web.service.ErrorResponse;
 import core.framework.internal.web.session.ReadOnlySession;
 import core.framework.internal.web.session.SessionManager;
 import core.framework.util.Strings;
+import core.framework.web.exception.ForbiddenException;
+import core.framework.web.exception.MethodNotAllowedException;
+import core.framework.web.exception.NotFoundException;
+import core.framework.web.exception.TooManyRequestsException;
 import core.framework.web.sse.ChannelListener;
 import io.undertow.server.HttpHandler;
 import io.undertow.server.HttpServerExchange;
 import io.undertow.util.HeaderMap;
 import io.undertow.util.Headers;
 import io.undertow.util.HttpString;
+import org.jspecify.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xnio.ChannelListeners;
@@ -98,11 +102,7 @@ public class ServerSentEventHandler implements HttpHandler {
         ChannelImpl<Object> channel = null;
         try {
             handlerContext.requestParser.parse(request, exchange, actionLog);
-            if (handlerContext.accessControl != null && !handlerContext.accessControl.allow(request.clientIP())) {
-                // check ip before checking routing, return 403 asap
-                exchange.setStatusCode(HTTPStatus.FORBIDDEN.code);
-                exchange.endExchange();
-            }
+            if (handlerContext.accessControl != null) handlerContext.accessControl.validate(request.clientIP());  // check ip before checking routing, return 403 asap
 
             actionLog.warningContext.maxProcessTimeInNano(MAX_PROCESS_TIME_IN_NANO);
             String path = request.path();
@@ -136,14 +136,21 @@ public class ServerSentEventHandler implements HttpHandler {
             support.listener.onConnect(request, channel, lastEventId);
 
             if (!channel.groups.isEmpty()) actionLog.context("group", channel.groups.toArray()); // may join group onConnect
+        } catch (NotFoundException | MethodNotAllowedException | ForbiddenException | TooManyRequestsException e) {
+            // do not treat those malicious requests as error w/ trace, to reduce overhead
+            // can still use errorHandler to render custom error page and potentially log error
+            handleError(e, channel, actionLog.id);
         } catch (Throwable e) {
             logManager.logError(e);
+            handleError(e, channel, actionLog.id);
+        }
+    }
 
-            if (channel != null) {
-                String message = errorMessage(handlerContext.responseBeanWriter.toJSON(ErrorResponse.errorResponse(e, actionLog.id)));
-                channel.sendBytes(Strings.bytes(message));
-                channel.close();    // gracefully shutdown connection to make sure retry/error can be sent
-            }
+    private void handleError(Throwable e, @Nullable ChannelImpl<Object> channel, String actionId) {
+        if (channel != null) {
+            String message = errorMessage(handlerContext.responseBeanWriter.toJSON(ErrorResponse.errorResponse(e, actionId)));
+            channel.sendBytes(Strings.bytes(message));
+            channel.close();    // gracefully shutdown connection to make sure retry/error can be sent
         }
     }